# docs/swagger.yaml
basePath: /
definitions:
...
  models.BankPaymentResponse:
    properties:
      amount:
        type: number
      created_at:
        type: string
      id:
        type: string
      paid_at:
        type: string
      payment_method:
        type: string
      return_id:
        type: string
      sale_id:
        type: string
      transaction_ref:
        type: string
      updated_at:
        type: string
    type: object
...
  models.CreateBankPaymentRequest:
    properties:
      amount:
        type: number
      payment_method:
        type: string
      return_id:
        type: string
      sale_id:
        type: string
    required:
    - amount
    - payment_method
    type: object
...
  models.OrderPaymentWebhook:
    properties:
      event_id:
        type: string
      event_type:
        type: string
      external_order_id:
        type: string
      paid_amount:
        type: number
      payment_date:
        type: string
      payment_method:
        type: string
      payment_status:
        type: string
      remarks:
        type: string
      timestamp:
        type: string
      transaction_id:
        type: string
    required:
    - event_id
    - event_type
    - external_order_id
    - paid_amount
    - payment_status
    - timestamp
    type: object
...
  models.PurchaseOrderItemResponse:
    properties:
      created_at:
        type: string
      id:
        type: string
      line_total:
        type: number
      po_id:
        type: string
      product_name:
        type: string
      product_sku:
        type: string
      quantity:
        type: integer
      received_quantity:
        type: integer
      unit_price:
        type: number
      variant_id:
        type: string
    type: object
  models.PurchaseOrderResponse:
    properties:
      actual_delivery_date:
        type: string
      collaborator_id:
        type: string
      collaborator_name:
        type: string
      created_at:
        type: string
      expected_delivery_date:
        type: string
      id:
        type: string
      items:
        items:
          $ref: '#/definitions/models.PurchaseOrderItemResponse'
        type: array
      order_date:
        type: string
      paid_amount:
        type: number
      payment_status:
        type: string
      po_number:
        type: string
      status:
        type: string
      total_amount:
        type: number
      updated_at:
        type: string
      warehouse_id:
        type: string
      warehouse_name:
        type: string
    type: object
...
  models.UpdatePOPaymentRequest:
    properties:
      paid_amount:
        type: number
      payment_status:
        description: unpaid, partial, paid
        type: string
    required:
    - paid_amount
    - payment_status
    type: object
...
  utils.ErrorResponseModel:
    properties:
      error:
        example: Detailed error message
        type: string
      message:
        example: Error occurred
        type: string
      success:
        example: false
        type: boolean
      timestamp:
        example: "2024-01-15T10:30:00Z"
        type: string
    type: object
  utils.Response:
    properties:
      data: {}
      error:
        type: string
      message:
        type: string
      success:
        type: boolean
      timestamp:
        type: string
    type: object
host: localhost:3000
info:
  contact:
    email: info@kisanlink.in
    name: Kisanlink Support
    url: https://github.com/Kisanlink/fpo-erp
  description: Comprehensive ERP system for agricultural cooperatives with multi-tenant
    architecture
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
  termsOfService: http://swagger.io/terms/
  title: Kisanlink ERP API
  version: "1.0"
paths:
...
  /api/v1/bank-payments:
    get:
      description: Retrieve all bank payments with pagination
      parameters:
      - description: 'Number of records to return (default: 10)'
        example: 10
        in: query
        name: limit
        type: integer
      - description: 'Number of records to skip (default: 0)'
        example: 0
        in: query
        name: offset
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Bank payments retrieved successfully
          schema:
            allOf:
            - $ref: '#/definitions/utils.Response'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/models.BankPaymentResponse'
                  type: array
              type: object
        "400":
          description: Bad request
          schema:
            $ref: '#/definitions/utils.ErrorResponseModel'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/utils.ErrorResponseModel'
        "403":
          description: Forbidden - insufficient permissions
          schema:
            $ref: '#/definitions/utils.ErrorResponseModel'
        "409":
          description: Conflict - resource already exists
          schema:
            $ref: '#/definitions/utils.ErrorResponseModel'
        "422":
          description: Unprocessable Entity - validation failed
          schema:
            $ref: '#/definitions/utils.ErrorResponseModel'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/utils.ErrorResponseModel'
      security:
      - BearerAuth: []
      summary: Get All Bank Payments
      tags:
      - Bank Payments
    post:
      consumes:
      - application/json
      description: Create a new bank payment record (requires authentication)
      parameters:
      - description: Bank payment data
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/models.CreateBankPaymentRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Bank payment created successfully
          schema:
            allOf:
            - $ref: '#/definitions/utils.Response'
            - properties:
                data:
                  $ref: '#/definitions/models.BankPaymentResponse'
              type: object
        "400":
          description: Bad request
          schema:
            $ref: '#/definitions/utils.ErrorResponseModel'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/utils.ErrorResponseModel'
        "403":
          description: Forbidden - insufficient permissions
          schema:
            $ref: '#/definitions/utils.ErrorResponseModel'
        "409":
          description: Conflict - resource already exists
          schema:
            $ref: '#/definitions/utils.ErrorResponseModel'
        "422":
          description: Unprocessable Entity - validation failed
          schema:
            $ref: '#/definitions/utils.ErrorResponseModel'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/utils.ErrorResponseModel'
      security:
      - BearerAuth: []
      summary: Create Bank Payment
      tags:
      - Bank Payments
  /api/v1/bank-payments/{id}:
    get:
      description: Retrieve a specific bank payment by ID
      parameters:
      - description: Payment ID
        example: PAY_12345678
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Bank payment retrieved successfully
          schema:
            allOf:
            - $ref: '#/definitions/utils.Response'
            - properties:
                data:
                  $ref: '#/definitions/models.BankPaymentResponse'
              type: object
        "400":
          description: Bad request
          schema:
            $ref: '#/definitions/utils.ErrorResponseModel'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/utils.ErrorResponseModel'
        "403":
          description: Forbidden - insufficient permissions
          schema:
            $ref: '#/definitions/utils.ErrorResponseModel'
        "404":
          description: Bank payment not found
          schema:
            $ref: '#/definitions/utils.ErrorResponseModel'
        "409":
          description: Conflict - resource already exists
          schema:
            $ref: '#/definitions/utils.ErrorResponseModel'
        "422":
          description: Unprocessable Entity - validation failed
          schema:
            $ref: '#/definitions/utils.ErrorResponseModel'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/utils.ErrorResponseModel'
      security:
      - BearerAuth: []
      summary: Get Bank Payment
      tags:
      - Bank Payments
  /api/v1/bank-payments/return/{returnID}:
    get:
      description: Retrieve all bank payments for a specific return
      parameters:
      - description: Return ID
        example: RET_12345678
        in: path
        name: returnID
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Bank payments for return retrieved successfully
          schema:
            allOf:
            - $ref: '#/definitions/utils.Response'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/models.BankPaymentResponse'
                  type: array
              type: object
        "400":
          description: Bad request
          schema:
            $ref: '#/definitions/utils.ErrorResponseModel'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/utils.ErrorResponseModel'
        "403":
          description: Forbidden - insufficient permissions
          schema:
            $ref: '#/definitions/utils.ErrorResponseModel'
        "409":
          description: Conflict - resource already exists
          schema:
            $ref: '#/definitions/utils.ErrorResponseModel'
        "422":
          description: Unprocessable Entity - validation failed
          schema:
            $ref: '#/definitions/utils.ErrorResponseModel'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/utils.ErrorResponseModel'
      security:
      - BearerAuth: []
      summary: Get Bank Payments by Return
      tags:
      - Bank Payments
  /api/v1/bank-payments/sale/{saleID}:
    get:
      description: Retrieve all bank payments for a specific sale
      parameters:
      - description: Sale ID
        example: SALE_12345678
        in: path
        name: saleID
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Bank payments for sale retrieved successfully
          schema:
            allOf:
            - $ref: '#/definitions/utils.Response'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/models.BankPaymentResponse'
                  type: array
              type: object
        "400":
          description: Bad request
          schema:
            $ref: '#/definitions/utils.ErrorResponseModel'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/utils.ErrorResponseModel'
        "403":
          description: Forbidden - insufficient permissions
          schema:
            $ref: '#/definitions/utils.ErrorResponseModel'
        "409":
          description: Conflict - resource already exists
          schema:
            $ref: '#/definitions/utils.ErrorResponseModel'
        "422":
          description: Unprocessable Entity - validation failed
          schema:
            $ref: '#/definitions/utils.ErrorResponseModel'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/utils.ErrorResponseModel'
      security:
      - BearerAuth: []
      summary: Get Bank Payments by Sale
      tags:
      - Bank Payments
...
  /api/v1/purchase-orders/{id}/payment:
    patch:
      consumes:
      - application/json
      description: Update the payment status of a purchase order (requires authentication)
      parameters:
      - description: 'Purchase Order ID (format: PORD_xxxxxxxx)'
        example: PORD_12345678
        in: path
        name: id
        required: true
        type: string
      - description: Payment update data
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/models.UpdatePOPaymentRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Payment status updated successfully
          schema:
            allOf:
            - $ref: '#/definitions/utils.Response'
            - properties:
                data:
                  $ref: '#/definitions/models.PurchaseOrderResponse'
              type: object
        "400":
          description: Bad request
          schema:
            $ref: '#/definitions/utils.ErrorResponseModel'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/utils.ErrorResponseModel'
        "403":
          description: Forbidden - insufficient permissions
          schema:
            $ref: '#/definitions/utils.ErrorResponseModel'
        "404":
          description: Purchase order not found
          schema:
            $ref: '#/definitions/utils.ErrorResponseModel'
        "409":
          description: Conflict - resource already exists
          schema:
            $ref: '#/definitions/utils.ErrorResponseModel'
        "422":
          description: Unprocessable Entity - validation failed
          schema:
            $ref: '#/definitions/utils.ErrorResponseModel'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/utils.ErrorResponseModel'
      security:
      - BearerAuth: []
      summary: Update Payment Status
      tags:
      - Purchase Orders
...
  /api/v1/webhooks/ecommerce/order/payment:
    post:
      consumes:
      - application/json
      description: Receives order.payment webhook from e-commerce platform
      parameters:
      - description: HMAC-SHA256 signature of request body
        in: header
        name: X-Webhook-Signature
        required: true
        type: string
      - description: Unique event identifier (for idempotency)
        in: header
        name: X-Webhook-Event-ID
        required: true
        type: string
      - description: Unix timestamp of webhook creation
        in: header
        name: X-Webhook-Timestamp
        required: true
        type: string
      - description: Order payment webhook payload
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/models.OrderPaymentWebhook'
      produces:
      - application/json
      responses:
        "200":
          description: Webhook processed successfully
          schema:
            allOf:
            - $ref: '#/definitions/utils.Response'
            - properties:
                data:
                  additionalProperties: true
                  type: object
              type: object
        "400":
          description: Bad request or invalid payload
          schema:
            $ref: '#/definitions/utils.ErrorResponseModel'
        "401":
          description: Invalid webhook signature
          schema:
            $ref: '#/definitions/utils.ErrorResponseModel'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/utils.ErrorResponseModel'
      summary: E-commerce Order Payment Webhook
      tags:
      - Webhooks
...
schemes:
- http
- https
securityDefinitions:
  BearerAuth:
    description: Type "Bearer" followed by a space and JWT token.
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"